fix(docs): autosummary templates for class methods
... and also attributes, exaptions or functions and attributes
on module level. When a class is added to an ".. autosummary"
directive, the docstring for the class itself is automatically
generated for the reference docs, but any methods or attributes
are not.

With a set of private autosummary templates and following the
comments in https://github.com/sphinx-doc/sphinx/issues/7912
now all elements will be respected by autosummary and all Python
dostrings will convert into RST.

Signed-off-by: Stephan Linz <linz@li-pro.net>